Refactor Session into a ZMQ-free base class.
This commit does not break API compatability.
